+++ litebird_sim/__init__.py
@@ -27,6 +27,19 @@ from .compress import (
     rle_compress,
     rle_decompress,
 )
+from .constants import (
+    ARCMIN_TO_RAD,
+    C_LIGHT_KM_S,
+    H_OVER_K_B,
+    T_CMB_K,
+    SOLAR_VELOCITY_KM_S,
+    SOLAR_VELOCITY_GAL_LAT_RAD,
+    SOLAR_VELOCITY_GAL_LON_RAD,
+    EARTH_L2_DISTANCE_KM,
+)
+from .units import (
+    Units,
+)
 from .coordinates import (
     DEFAULT_COORDINATE_SYSTEM,
     DEFAULT_TIME_SCALE,
@@ -207,6 +220,17 @@ __all__ = [
     "allocate_alm",
     "gauss_beam_to_alm",
     "generate_gauss_beam_alms",
+    # constants.py
+    "ARCMIN_TO_RAD",
+    "C_LIGHT_KM_S",
+    "H_OVER_K_B",
+    "T_CMB_K",
+    "SOLAR_VELOCITY_KM_S",
+    "SOLAR_VELOCITY_GAL_LAT_RAD",
+    "SOLAR_VELOCITY_GAL_LON_RAD",
+    "EARTH_L2_DISTANCE_KM",
+    # units.py
+    "Units",
     # compress.py
     "rle_compress",
     "rle_decompress",